Added file of my genus computations
This file details the calculated genus, running time, etc from running this algorithm over the first 3000 triangulations in the Hodgson-Weeks closed hyperbolic census.
As triangulations in this census are not necessarily simplified, 'census isosig' refers to the isomorphism signature as the triangulation appears in the census. I have simplified triangulations where necessary. Additionally, some may have had the gadget inserted, so 'final isosig' represents the isomorphism signature of the triangulation in the final successful calculation.